Moderator request for subclub - Assign mod random
Members can request to be a moderator of a subclub, also admin can view the mod-requests of a subclub and assign a mod to a subclub randomly. "make-moderator" and "mod-requests" are unnecessary for the final product. However, they are helpful for development now. Lastly, a member who was banned when s/he was a moderator cannot send mod-request.